Fix build break Document Events classes
