+++ tests/src/Interop/COM/ServerContracts/Server.Events.cs
@@ -4,7 +4,7 @@
 
 #pragma warning disable 618 // Must test deprecated features
 
-namespace Server.Contract.Events
+namespace Server.Contract
 {
     using System;
     using System.Collections.Generic;
@@ -13,131 +13,178 @@ namespace Server.Contract.Events
     using IConnectionPoint = System.Runtime.InteropServices.ComTypes.IConnectionPoint;
     using IConnectionPointContainer = System.Runtime.InteropServices.ComTypes.IConnectionPointContainer;
 
-    [ComVisible(false)]
-    public delegate void TestingEvents_OnEventEventHandler(string msg);
-
-    [ComVisible(false)]
-    [ComEventInterface(typeof(Contract.TestingEvents), typeof(TestingEvents_EventProvider))]
-    public interface TestingEvents_Event
+    namespace Servers
     {
-        event TestingEvents_OnEventEventHandler OnEvent;
-    }
-
-    public sealed class TestingEvents_EventProvider : TestingEvents_Event, IDisposable
-    {
-        private readonly WeakReference ConnectionPointContainer;
-        private readonly List<TestingEvents_SinkHelper> eventSinkHelpers = new List<TestingEvents_SinkHelper>();
-
-        private IConnectionPoint connectionPoint;
-        private bool isDisposed = false;
-
-        public TestingEvents_EventProvider(object container)
+        /// <summary>
+        /// Managed definition of CoClass
+        /// </summary>
+        [ComImport]
+        [CoClass(typeof(EventTestingClass))]
+        [Guid("83AFF8E4-C46A-45DB-9D91-2ADB5164545E")]
+        internal interface EventTesting : IEventTesting, Events.TestingEvents_Event
         {
-            this.ConnectionPointContainer = new WeakReference((IConnectionPointContainer)container, false);
         }
 
-        event TestingEvents_OnEventEventHandler TestingEvents_Event.OnEvent
+        /// <summary>
+        /// Managed activation for CoClass
+        /// </summary>
+        [ComImport]
+        [ComSourceInterfaces("Server.Contract.Events.TestingEvents\0")]
+        [Guid(Server.Contract.Guids.EventTesting)]
+        internal class EventTestingClass
         {
-            add
+        }
+    }
+
+    /// <summary>
+    /// Classes in the Events namespace are traditionally generated by the TlbImp tool.
+    /// </summary>
+    namespace Events
+    {
+        /// <summary>
+        /// Delegate used for event handler
+        /// </summary>
+        [ComVisible(false)]
+        public delegate void TestingEvents_OnEventEventHandler(string msg);
+
+        /// <summary>
+        /// Event source interface
+        /// </summary>
+        /// <remarks>
+        /// Observe usage of the <see cref="ComEventInterfaceAttribute"/> attribute.
+        /// </remarks>
+        [ComVisible(false)]
+        [ComEventInterface(typeof(Contract.TestingEvents), typeof(TestingEvents_EventProvider))]
+        public interface TestingEvents_Event
+        {
+            event TestingEvents_OnEventEventHandler OnEvent;
+        }
+
+        /// <summary>
+        /// Managed proxy for event subscription via IConnectionPointContainer and IConnectionPoint.
+        /// </summary>
+        public sealed class TestingEvents_EventProvider : TestingEvents_Event, IDisposable
+        {
+            private readonly WeakReference ConnectionPointContainer;
+            private readonly List<TestingEvents_SinkHelper> eventSinkHelpers = new List<TestingEvents_SinkHelper>();
+
+            private IConnectionPoint connectionPoint;
+            private bool isDisposed = false;
+
+            public TestingEvents_EventProvider(object container)
             {
-                lock (this.eventSinkHelpers)
-                {
-                    if (this.connectionPoint == null)
-                    {
-                        this.Init();
-                    }
-
-                    var sinkHelper = new TestingEvents_SinkHelper();
-
-                    int cookie;
-                    this.connectionPoint.Advise(sinkHelper, out cookie);
-
-                    sinkHelper.Cookie = cookie;
-                    sinkHelper.OnEventDelegate = value;
-                    this.eventSinkHelpers.Add(sinkHelper);
-                }
+                this.ConnectionPointContainer = new WeakReference((IConnectionPointContainer)container, false);
             }
-            remove
+
+            event TestingEvents_OnEventEventHandler TestingEvents_Event.OnEvent
             {
-                lock (this.eventSinkHelpers)
+                add
                 {
-                    TestingEvents_SinkHelper sinkHelper = null;
-                    int removeIdx = -1;
-                    for (int i = 0; i < this.eventSinkHelpers.Count; ++i)
+                    lock (this.eventSinkHelpers)
                     {
-                        TestingEvents_SinkHelper sinkHelperMaybe = this.eventSinkHelpers[i];
-                        if (sinkHelperMaybe.OnEventDelegate.Equals(value))
+                        if (this.connectionPoint == null)
                         {
-                            removeIdx = i;
-                            sinkHelper = sinkHelperMaybe;
-                            break;
+                            this.Init();
+                        }
+
+                        var sinkHelper = new TestingEvents_SinkHelper();
+
+                        int cookie;
+                        this.connectionPoint.Advise(sinkHelper, out cookie);
+
+                        sinkHelper.Cookie = cookie;
+                        sinkHelper.OnEventDelegate = value;
+                        this.eventSinkHelpers.Add(sinkHelper);
+                    }
+                }
+                remove
+                {
+                    lock (this.eventSinkHelpers)
+                    {
+                        TestingEvents_SinkHelper sinkHelper = null;
+                        int removeIdx = -1;
+                        for (int i = 0; i < this.eventSinkHelpers.Count; ++i)
+                        {
+                            TestingEvents_SinkHelper sinkHelperMaybe = this.eventSinkHelpers[i];
+                            if (sinkHelperMaybe.OnEventDelegate.Equals(value))
+                            {
+                                removeIdx = i;
+                                sinkHelper = sinkHelperMaybe;
+                                break;
+                            }
+                        }
+
+                        if (removeIdx < 0)
+                        {
+                            return;
+                        }
+
+                        this.connectionPoint.Unadvise(sinkHelper.Cookie);
+                        this.eventSinkHelpers.RemoveAt(removeIdx);
+
+                        if (this.eventSinkHelpers.Count == 0)
+                        {
+                            Marshal.ReleaseComObject(this.connectionPoint);
+                            this.connectionPoint = null;
                         }
                     }
-
-                    if (removeIdx < 0)
-                    {
-                        return;
-                    }
-
-                    this.connectionPoint.Unadvise(sinkHelper.Cookie);
-                    this.eventSinkHelpers.RemoveAt(removeIdx);
-
-                    if (this.eventSinkHelpers.Count == 0)
-                    {
-                        Marshal.ReleaseComObject(this.connectionPoint);
-                        this.connectionPoint = null;
-                    }
                 }
             }
-        }
 
-        void IDisposable.Dispose()
-        {
-            if (this.isDisposed)
+            void IDisposable.Dispose()
             {
-                return;
-            }
-
-            lock (this.eventSinkHelpers)
-            {
-                foreach (TestingEvents_SinkHelper sinkHelper in this.eventSinkHelpers)
+                if (this.isDisposed)
                 {
-                    this.connectionPoint.Unadvise(sinkHelper.Cookie);
+                    return;
                 }
 
-                this.eventSinkHelpers.Clear();
+                lock (this.eventSinkHelpers)
+                {
+                    foreach (TestingEvents_SinkHelper sinkHelper in this.eventSinkHelpers)
+                    {
+                        this.connectionPoint.Unadvise(sinkHelper.Cookie);
+                    }
+
+                    this.eventSinkHelpers.Clear();
+                }
+
+                Marshal.ReleaseComObject(this.connectionPoint);
+                this.connectionPoint = null;
+
+                this.isDisposed = true;
+                System.GC.SuppressFinalize(this);
             }
 
-            Marshal.ReleaseComObject(this.connectionPoint);
-            this.connectionPoint = null;
-
-            this.isDisposed = true;
-            System.GC.SuppressFinalize(this);
-        }
-
-        private void Init()
-        {
-            var container = (IConnectionPointContainer)this.ConnectionPointContainer.Target;
-
-            Guid iid = typeof(Contract.TestingEvents).GUID;
-            IConnectionPoint connectionPoint;
-            container.FindConnectionPoint(ref iid, out connectionPoint);
-
-            this.connectionPoint = connectionPoint;
-        }
-    }
-
-    [ClassInterface(ClassInterfaceType.None)]
-    public class TestingEvents_SinkHelper : Contract.TestingEvents
-    {
-        public int Cookie { get; set; }
-        public TestingEvents_OnEventEventHandler OnEventDelegate { get; set; }
-
-        public void OnEvent(string msg)
-        {
-            if (this.OnEventDelegate != null)
+            private void Init()
             {
-                this.OnEventDelegate(msg);
+                var container = (IConnectionPointContainer)this.ConnectionPointContainer.Target;
+
+                Guid iid = typeof(Contract.TestingEvents).GUID;
+                IConnectionPoint connectionPoint;
+                container.FindConnectionPoint(ref iid, out connectionPoint);
+
+                this.connectionPoint = connectionPoint;
+            }
+        }
+
+        /// <summary>
+        /// Wrapper for event delegate.
+        /// </summary>
+        /// <remarks>
+        /// Observe usage of the <see cref="ClassInterfaceAttribute"/> attribute.
+        /// </remarks>
+        [ClassInterface(ClassInterfaceType.None)]
+        public class TestingEvents_SinkHelper : Contract.TestingEvents
+        {
+            public int Cookie { get; set; }
+            public TestingEvents_OnEventEventHandler OnEventDelegate { get; set; }
+
+            public void OnEvent(string msg)
+            {
+                if (this.OnEventDelegate != null)
+                {
+                    this.OnEventDelegate(msg);
+                }
             }
         }
     }
